Progress on re-coding universe data structure
Currently, the universe is defined using Nodes in the hidden
scene. This is starting to not make sense as none of the
features of the Nodes are being utilized besides its tree data
structure. Actual data related to space, such as position and
planet parameters, are stored as Components: AstronomicalBody
and Entity, Satellite being an abstract base class with absolute
position. This is unnecessarily complex in inflexible,
implemented with the thought that interfacing with the existing
Node class would be simpler.

Also, ActiveArea manages a single PlanetTerrain, meaning that
only a single planet can be visible at once. ActiveArea
shouldn't care about what PlanetTerrain does at all.

This commit aims to do the following:

* Separate Satellite into its own tree data structure
* Use relative long coordinates
* Make PlanetTerrain independent of ActiveArea
* Make ActiveArea into a Satellite

Many features related to planets and space are not functional
for the time being.