Implement ActiveArea distance loading
The ActiveArea Satellite doesn't move yet, but it now loops
through every Satellite, and loads them based on their relative
distance. Satellites have an imaginary sphere around them
defined by its new m_loadRadius variable. When this sphere
intersects an ActiveArea's sphere, then the ActiveArea will
attempt to load it.

ActiveArea needs a satellite to follow around, so the NodeSat
Satellite class was created to position nodes, such as the
player's craft, in the large universe. NodeSat keeps a SharedPtr
of a Node, which it can add to an ActiveArea scene when loaded.

Notable changes:

* Added NodeSat class, with headers and source files
* Updated Satellite class
  * Added m_index, index in parent's m_children array
  * Added getter and setter functions
  * Added float position argument for load function
* Updated OspUniverse
  * Commented out load functions
  * Added a NodeSat instance for the player's craft
* Updated AstronomicalBody inheriting changes from Satellite
* Updated ActiveArea
  * Implemented recursive loop through all Satellites,
    simultaneously calculating relative positions
  * Added distance_check_then_load function
* A few other changes