Set Object as the base class of Satellite
Making Satellite an Object allows it to receive Urho3D event
handlers and have type identification, while still having the
same reference counting. This was done mainly because of a
diamond problem with ActiveArea, inheriting RefCounted twice
through LogicComponent and Satellite. This caused the crash
described in commit (5ff140a). Inheritance to LogicComponent was
removed, and Satellite was changed to an Object so that
ActiveArea can receive to events from the scene.

Notable changes:

* Updated Satellite class
  * Base class changed to Object, and updated it and all of it's
    derived classes
  * Replaced m_unitsPerMeter with m_precision, see comments
  * Added is_loaded() function
  * Added calculate_position() function, see comments
  * Implemented load() for derived classes to call
* A few other changes