Clean PlanetWrenderer again
While trying to just get updating to work again, code related to
the base subdividing icosahedron was separated into the
IcoSphereTree class, and everything related to drawing it was
removed entirely. This was done for readability, support for
multiple ActiveAreas on the same planet, multiplayer, and
terrain collision. Still, everything here is half-implemented.